requirements.txt
robotframework==6.1.1
robotframework-appiumlibrary==2.0.0
pyyaml==6.0.1
robotframework-faker==5.0.0
robotframework-requests==0.9.6
robotframework-pabot==2.16.0
oracledb
Appium-Python-Client==3.1.1;python_version >= '3.0'